Made error printing more visible
I found myself not having seen error conditions, usually when the cookie
is not good anymore, i.e. due to IP changes. Now the error output is
more distinct and at least for me, works better.

An idea might be to have colourful output if a capable terminal was
detected.